backend/requirements.txt
email-validator==2.3.0
fastapi==0.115.8
pydantic-settings==2.7.1
sqlmodel==0.0.22
uvicorn[standard]==0.34.0
